+++ packages/media-parser/src/iterator/buffer-iterator.ts
@@ -5,31 +5,16 @@ import {
 	knownIdsWithTwoLength,
 } from '../containers/webm/segments/all-segments';
 import {detectFileType} from '../file-types';
-import type {ParseMediaMode} from '../options';
+import {bufferManager} from './buffer-manager';
 import {makeOffsetCounter} from './offset-counter';
 
 export const getArrayBufferIterator = (
 	initialData: Uint8Array,
 	maxBytes: number | null,
 ) => {
-	const buf = new ArrayBuffer(initialData.byteLength, {
-		maxByteLength:
-			maxBytes === null
-				? initialData.byteLength
-				: Math.min(maxBytes as number, 2 ** 32),
-	});
-	if (!buf.resize) {
-		throw new Error(
-			'`ArrayBuffer.resize` is not supported in this Runtime. On the server: Use at least Node.js 20 or Bun. In the browser: Chrome 111, Edge 111, Safari 16.4, Firefox 128, Opera 111',
-		);
-	}
-
-	let uintArray = new Uint8Array(buf);
-
-	uintArray.set(initialData);
-
-	let view = new DataView(uintArray.buffer);
 	const counter = makeOffsetCounter(0);
+	const {uintArray, view, addData, destroy, removeBytesRead, skipTo} =
+		bufferManager({initialData, maxBytes, counter});
 
 	const startCheckpoint = () => {
 		const checkpoint = counter.getOffset();
@@ -217,68 +202,8 @@ export const getArrayBufferIterator = (
 		return val;
 	};
 
-	const addData = (newData: Uint8Array) => {
-		const oldLength = buf.byteLength;
-		const newLength = oldLength + newData.byteLength;
-		if (newLength < oldLength) {
-			throw new Error('Cannot decrement size');
-		}
-
-		if (newLength > (maxBytes ?? Infinity)) {
-			throw new Error(
-				`Exceeded maximum byte length ${maxBytes} with ${newLength}`,
-			);
-		}
-
-		buf.resize(newLength);
-		const newArray = new Uint8Array(buf);
-		newArray.set(newData, oldLength);
-		uintArray = newArray;
-		view = new DataView(uintArray.buffer);
-	};
-
 	const bytesRemaining = () => {
 		return uintArray.byteLength - counter.getDiscardedOffset();
-	};
-
-	const removeBytesRead = (force: boolean, mode: ParseMediaMode) => {
-		const bytesToRemove = counter.getDiscardedOffset();
-
-		// Only do this operation if it is really worth it 😇
-		// let's set the threshold to 3MB
-		if (bytesToRemove < 3_000_000 && !force) {
-			return {bytesRemoved: 0, removedData: null};
-		}
-
-		// Don't remove if the data is not even available
-		if (view.byteLength < bytesToRemove && !force) {
-			return {bytesRemoved: 0, removedData: null};
-		}
-
-		counter.discardBytes(bytesToRemove);
-
-		const removedData =
-			mode === 'download' ? uintArray.slice(0, bytesToRemove) : null;
-
-		const newData = uintArray.slice(bytesToRemove);
-		uintArray.set(newData);
-		buf.resize(newData.byteLength);
-		view = new DataView(uintArray.buffer);
-
-		return {bytesRemoved: bytesToRemove, removedData};
-	};
-
-	const skipTo = (offset: number) => {
-		const becomesSmaller = offset < counter.getOffset();
-		if (!becomesSmaller) {
-			const currentOffset = counter.getOffset();
-			counter.increment(offset - currentOffset);
-			return;
-		}
-
-		buf.resize(0);
-		counter.decrement(counter.getOffset() - offset);
-		counter.setDiscardedOffset(offset);
 	};
 
 	const readExpGolomb = () => {
@@ -408,11 +333,6 @@ export const getArrayBufferIterator = (
 		}
 
 		return result;
-	};
-
-	const destroy = () => {
-		uintArray = new Uint8Array(0);
-		buf.resize(0);
 	};
 
 	return {
