+++ packages/media-parser/src/iterator/offset-counter.ts
@@ -28,3 +28,5 @@ export const makeOffsetCounter = (initial: number) => {
 		},
 	};
 };
+
+export type OffsetCounter = ReturnType<typeof makeOffsetCounter>;
